move balance tree to dedicated crate
